requirements.txt
virtualenv
django
django-scheduler
